refactor. components 폴더 내 컴포넌트 파일 페이지 폴더로 분리
