+++ src/components/cart/CartItem.tsx
@@ -1,20 +1,20 @@
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import Button from '../elements/Button';
-import useViewport from '../hooks/useViewport';
-import usePutCartQuery from '../query/usePutCartQuery';
-import type { TCartList } from '../redux/reducer/cartSlice';
+import Button from '../../elements/Button';
+import useViewport from '../../hooks/useViewport';
+import usePutCartQuery from '../../query/usePutCartQuery';
+import type { TCartList } from '../../redux/reducer/cartSlice';
 import {
   setCheckedList,
   setKeywordNo,
   setProductNo,
-} from '../redux/reducer/cartSlice';
-import { useAppDispatch } from '../redux/store';
-import theme from '../shared/style/theme';
-import type { IProductPayCart, TUserOption } from '../shared/types/types';
-import * as t from '../style/cartItem.style';
+} from '../../redux/reducer/cartSlice';
+import { useAppDispatch } from '../../redux/store';
+import theme from '../../shared/style/theme';
+import type { IProductPayCart, TUserOption } from '../../shared/types/types';
+import * as t from '../../style/cartItem.style';
+import GlobalModal from '../common/GlobalModal';
 import CartOptionModal from './CartOptionModal';
-import GlobalModal from './common/GlobalModal';
 
 interface IProps {
   cartList: IProductPayCart[];
